Extract graphql factory into its own submodule
The GraphQL API is now becoming the dominant API to use with graphgen,
but it is all blopped into a single file. This makes it difficult to
work with individual pieces of it, but since we need to expose more
metadata about the GraphQL analysis for the inspector tool, this is
difficult.

To make working with the individual pieces of the GraphQL factory
generator, this moves the entire graphql subsystem into its own
directory, and then the individual `types.ts` are exported from
there.

In the future, we should probably move the graph storage piece into
its own directory.
