+++ mod.ts
@@ -1,4 +1,5 @@
 export * from "./src/graph.ts";
-export * from "./src/graphql.ts";
+export * from "./src/graphql/graphgen.ts";
+export * from "./src/graphql/types.ts";
 export * from "./src/seedrandom.ts";
 export * from "./src/view.ts";
